+++ WithSecure/withsecure/models.py
@@ -1,5 +1,6 @@
-from pydantic import BaseModel, Field
 from typing import Any
+
+from pydantic import BaseModel, Field
 
 
 class WithSecureModuleConfiguration(BaseModel):